Break out common cloud functionality to clouds/base.py
